RAC-326 refactor : 결제 리팩토링
