+++ src/main/java/com/postgraduate/domain/payment/application/usecase/PaymentManageUseCase.java
@@ -13,7 +13,6 @@ import com.postgraduate.domain.payment.exception.RefundFailException;
 import com.postgraduate.domain.senior.domain.entity.Senior;
 import com.postgraduate.domain.senior.domain.service.SeniorGetService;
 import com.postgraduate.domain.user.domain.entity.User;
-import com.postgraduate.domain.user.domain.entity.constant.Role;
 import com.postgraduate.domain.user.domain.service.UserGetService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -28,7 +27,7 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
 
-import static com.postgraduate.domain.payment.application.usecase.PaymentParameter.*;
+import static com.postgraduate.domain.payment.application.usecase.constant.PaymentParameter.*;
 import static com.postgraduate.domain.payment.presentation.constant.PaymentResponseMessage.FAIL_PAYMENT;
 import static com.postgraduate.domain.user.domain.entity.constant.Role.ADMIN;
 import static org.springframework.http.CacheControl.noCache;
