+++ src/main/java/com/postgraduate/domain/payment/application/usecase/constant/PaymentParameter.java
@@ -1,4 +1,4 @@
-package com.postgraduate.domain.payment.application.usecase;
+package com.postgraduate.domain.payment.application.usecase.constant;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
